+++ Android/app/src/main/java/bitc/fullstack502/final_project_team1/network/dto/SurveyResultDetailDto.kt
@@ -21,7 +21,7 @@ data class SurveyResultDetailDto(
     val intEditPhoto: String?,
     val status: String?,
     val buildingId: Long?,
-    val userId: Int?,
+    val userId: Long?,
     val createdAt: String?,
     val updatedAt: String?
 )
